Rewrite and restructure the README
Improves the README to have better structure and ordering for readers,
more examples, and clearer explanations. It aligns the README with the
structure of many other modern GitHub-based utilities' READMEs.

Not just tradition, these conventions make a lot of sense:

* There's a quick demonstration of the interface at the top of the file,
  allowing visitors to quickly examine the value and design of the tool.

* Installation instructions are near the top.

* There are lots of examples (with syntax higlighting).

* Commands' documentation is structured and grouped together.

* Lots of links to external resources.

The commands for fixing Ubuntu 12.04 installations was replaced with a
link to the PPA.Ubuntu users should know how to install PPAs, but if
not, the Meteorite README probably isn't the best place to learn.
There's no need to bother non-Ubuntu users with those instructions.
Also, the PPA isn't the only way to install a recent version of Node,
and so it shouldn't be presented as such.