Create "contributing" document from wiki contents
Create a `CONTRIBUTING.md` document from the "Contributing" and
"Testing" pages on the GitHub wiki.

Removed the link to the "Contributing" page on the README.

I'm not involved with Meteorite development, so the notes on the wiki
may be terribly out of date. Even so, I thought I'd include them if need
be. They can be fixed up or removed later.