Revert (to some degree) e7003078f5638680e73495bf2fe1aa7506f28424
When using the "latest" version of a package, I stored the version 
(from atmosphere) on the package, so that I could use it on logging install.

This was a bit of a hack and it burnt me in make the smart.lock write incorrectly (there's no way to tell at that point that we are on the latest).

The correct solution is to not send a version when installing the latest and letting atmosphere figure it out and log it.